massive data formats change

+++ DataFormats/interface/NTEvent.h
@@ -9,43 +9,43 @@ namespace ztop{
     explicit NTEvent(){}
     ~NTEvent(){}
     //sets
-    void setRunNo(double RunNo){runNo_=RunNo;}
-    void setLumiBlock(double LumiBl){lumiBl_=LumiBl;}
-    void setEventNo(double EventNo){eventNo_=EventNo;}
+    void setRunNo(float RunNo){runNo_=RunNo;}
+    void setLumiBlock(float LumiBl){lumiBl_=LumiBl;}
+    void setEventNo(float EventNo){eventNo_=EventNo;}
 
-    void setVertexMulti(double VertexMulti){vertexMulti_=VertexMulti;}
+    void setVertexMulti(float VertexMulti){vertexMulti_=VertexMulti;}
     void setFiredTriggers(std::vector<std::string> Triggers){triggers_=Triggers;}
 
     void setTruePU(float BXminus, float BXzero, float BXplus){truePU_[0]=BXminus; truePU_[1]=BXzero; truePU_[2]=BXplus;}
 
-    void setIsoRho(std::vector<double> Rhos){rhos_=Rhos;}
-    void setPDFWeights(std::vector<double> Weights){pdfweights_=Weights;}
+    void setIsoRho(std::vector<float> Rhos){rhos_=Rhos;}
+    void setPDFWeights(std::vector<float> Weights){pdfweights_=Weights;}
 
     //gets
-    double runNo(){return runNo_;}
-    double lumiBlock(){return lumiBl_;}
-    double eventNo(){return eventNo_;}
+    float runNo(){return runNo_;}
+    float lumiBlock(){return lumiBl_;}
+    float eventNo(){return eventNo_;}
 
-    double vertexMulti(){return vertexMulti_;}
+    float vertexMulti(){return vertexMulti_;}
    const std::vector<std::string>& firedTriggers()const{return triggers_;}
 
     const float& truePU(int BX=0)const{return truePU_[1+BX];}
-    double trueAvgPU()const{return (truePU_[0]+truePU_[1]+truePU_[2])/3;}
+    float trueAvgPU()const{return (truePU_[0]+truePU_[1]+truePU_[2])/3;}
 
-    double isoRho(unsigned int identifier){return rhos_[identifier];}
-    const double& PDFWeight(unsigned int id)const{return pdfweights_[id];}
+    float isoRho(unsigned int identifier){return rhos_[identifier];}
+    const float& PDFWeight(unsigned int id)const{return pdfweights_[id];}
     unsigned int PDFWeightsSize(){return pdfweights_.size();}
 
   protected:
-    double runNo_;
-    double lumiBl_;
-    double eventNo_;
+    float runNo_;
+    float lumiBl_;
+    float eventNo_;
 
-    double vertexMulti_;
+    float vertexMulti_;
     std::vector<std::string> triggers_;
 
-    std::vector<double> rhos_;
-    std::vector<double> pdfweights_;
+    std::vector<float> rhos_;
+    std::vector<float> pdfweights_;
 
     float truePU_[3];
