+++ DataFormats/interface/NTIsolation.h
@@ -8,23 +8,23 @@ namespace ztop{
     explicit NTIsolation(){ChargedHadronIso_=-999;NeutralHadronIso_=-999;PhotonIso_=-999;PuChargedHadronIso_=-999;}
     ~NTIsolation(){}
     //sets
-    void setChargedHadronIso(double ChargedHadronIso){ChargedHadronIso_=ChargedHadronIso;}
-    void setNeutralHadronIso(double NeutralHadronIso){NeutralHadronIso_=NeutralHadronIso;}
-    void setPhotonIso(double PhotonIso){PhotonIso_=PhotonIso;}
-    void setPuChargedHadronIso(double PuChargedHadronIso){PuChargedHadronIso_=PuChargedHadronIso;}
+    void setChargedHadronIso(float ChargedHadronIso){ChargedHadronIso_=ChargedHadronIso;}
+    void setNeutralHadronIso(float NeutralHadronIso){NeutralHadronIso_=NeutralHadronIso;}
+    void setPhotonIso(float PhotonIso){PhotonIso_=PhotonIso;}
+    void setPuChargedHadronIso(float PuChargedHadronIso){PuChargedHadronIso_=PuChargedHadronIso;}
 
 
     //gets
-    const double &chargedHadronIso()const{return ChargedHadronIso_;}
-    const double &neutralHadronIso()const{return NeutralHadronIso_;}
-    const double &photonIso()const{return PhotonIso_;}
-    const double &puChargedHadronIso()const{return PuChargedHadronIso_;}
+    const float &chargedHadronIso()const{return ChargedHadronIso_;}
+    const float &neutralHadronIso()const{return NeutralHadronIso_;}
+    const float &photonIso()const{return PhotonIso_;}
+    const float &puChargedHadronIso()const{return PuChargedHadronIso_;}
 
   protected:
-    double ChargedHadronIso_;
-    double NeutralHadronIso_;
-    double PhotonIso_;
-    double PuChargedHadronIso_;
+    float ChargedHadronIso_;
+    float NeutralHadronIso_;
+    float PhotonIso_;
+    float PuChargedHadronIso_;
 
   };
 }
